+++ packages/expo-router/plugin/src/withRouter.ts
@@ -99,8 +99,6 @@ export type Props = {
    * @deprecated
    */
   unstable_useServerMiddleware?: boolean;
-  /** Enable experimental data loader support. Requires `web.output: 'static' | 'server'` to be set in app config. */
-  unstable_useServerDataLoaders?: boolean;
   /** Enable experimental server-side rendering. When enabled with `web.output: 'server'`, HTML is rendered at request time instead of being pre-rendered at build time. */
   unstable_useServerRendering?: boolean;
   /** Disable synchronous layout updates for native screens. */